move test to StoreTest
